requirements.txt
coreapi==2.3.3
coreschema==0.0.4
djangorestframework==3.12.2
asgiref==3.4.1
Django==3.1.5
sqlparse==0.4.2
tzdata==2021.5
